refactor views.py > accounts

+++ train_ecommerce/accounts/views.py
@@ -6,14 +6,18 @@ from django.shortcuts import render, redirect
 from django.views import View
 from .forms import *
 from .models import *
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from datetime import timedelta, datetime
 
 
 class UserRegisterView(View):
     form_class = UserRegistrationForm
+    template_name = 'accounts/register.html'
 
     def get(self, request):
         form = self.form_class
-        return render(request, 'accounts/register.html', {'form': form})
+        return render(request, self.template_name, {'form': form})
 
     def post(self, request):
         form = self.form_class(request.POST)
@@ -31,31 +35,65 @@ class UserRegisterView(View):
             }
             messages.success(request, 'we sent you a code', 'success')
             return redirect('accounts:verify_code')
-        return redirect('home:home')
+        return render(request, self.template_name, {'form': form})
 
 
 class UserRegisterVerifyCodeView(View):
     form_class = VerifyCodeForm
+    template_name = 'accounts/verify_code.html'
 
     def get(self, request):
         form = self.form_class
-        return render(request, 'accounts/verify_code.html', {'form': form})
+        return render(request, self.template_name, {'form': form})
 
     def post(self, request):
         user_session = request.session['user_registration_info']
-        code_instance = OtmCode.objects.get(phone_number=user_session['phone'])
+        code_instance = OtpCode.objects.get(phone_number=user_session['phone_number'])
         form = self.form_class(request.POST)
 
         if form.is_valid():
             cd = form.cleaned_data
 
             if cd['code'] == code_instance.code:
-                user = User.objects.create_user(user_session['phone_number'], user_session['email'],
-                                                user_session['full_name'], user_session['password'])
-                code_instance.delete()
-                messages.success(request, 'registered', 'success')
+                now = datetime.now()
+                if (code_instance.created + timedelta(minutes=1)) < now:
+                    user = User.objects.create_user(user_session['phone_number'], user_session['email'],
+                                                    user_session['full_name'], user_session['password'])
+                    code_instance.delete()
+                    messages.success(request, 'registered', 'success')
+                else:
+                    messages.error(request, 'time is up!', 'danger')
+                    return redirect(self.template_name)
             else:
                 messages.error(request, 'wrong code!', 'danger')
-                return redirect('accounts:verify_code')
+                return redirect(self.template_name)
 
+        return redirect('accounts:user_register')
+
+
+class UserLoginView(View):
+    form_class = UserLoginForm
+    template_name = 'accounts/login.html'
+
+    def get(self, request):
+        form = self.form_class
+        return render(request, self.template_name, {'form': form})
+
+    def post(self, request):
+        form = self.form_class(request.POST)
+        if form.is_valid():
+            cd = form.cleaned_data
+            user = authenticate(request, phone_number=cd['phone'], password=cd['password'])
+            login(request, user)
+            request.session['user_login_info'] = {
+                'phone_number': cd['phone'],
+                'password': cd['password'],
+            }
+            return redirect('home:home')
+        return render(request, self.template_name, {'form': form})
+
+
+class UserLogoutView(View):
+    def get(self, request):
+        logout(request)
         return redirect('home:home')
